Day 3 part 2 jhfc that was hard.
Building the tree is O(M*N) for M rows of N bits each, then we just need to walk the tree
once for the O2 and once for the CO2, each walk is O(N). So the whole things is like
O((M+2)N) (I don't know if that's a thing.